Added TODOs for making logic for singledeposit, withdraw & swap to work properly
